100 show disc number on album page (#103)
* Added disc number to the album page. Also fixed a bug where updating an album was adding an album in the local story.

* Undoing some rename shenanigans

* Fixed an issue where the page filters would go hidden if there were on search results.
Also, fixed a rows miscount issue on the songs catalogue.
Then also fixed the song download button on the albums screen.

* updated python and setup service roles to add group write access so some generated files in our python library can be deleted.

* Had to update the rest of the python-setup include roles to pass in the group

*
 --Refactored logging.
--Updated the startup api playbook to always restart the api service.
--Added an error log to the job service

+++ src/musical_chairs_libs/dtos_and_utilities/queued_item.py
@@ -44,7 +44,7 @@ class SongListDisplayItem(QueuedItem):
 	artist: Optional[str]
 	path: str
 	internalpath: str
-	track: Optional[int]=None
+	track: Optional[float]=None
 	playedtimestamp: Optional[float]=None
 	rules: list[ActionRule]=cast(
 		list[ActionRule],
